feat/improvement: added ReductionPhases, next_cigarette based on phase; refactored code
